Refactor: null 검사 이동

+++ src/main/java/com/flash21/caddycom/service/golfFieldDetail/CommentService.java
@@ -63,6 +63,7 @@ public class CommentService {
      */
     @Transactional
     public void deleteComments(List<Long> commentIds) {
+        if(commentIds == null || commentIds.isEmpty()) return;
         commentRepository.deleteAllByIdInBatch(commentIds);
     }
